pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "latent-link-gateway"
version = "0.1.0"
description = "Latent-Link Rule Injection Gateway - compress engineering rules into latent space tensors via LatentMAS"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.44.0",
    "safetensors>=0.4.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "httpx>=0.27.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio", "ruff"]
quantize = ["onnxruntime>=1.16.0", "optimum>=1.16.0"]

[project.scripts]
latent-compile = "src.compiler.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
